Fix Rack warning about unprocessable_entity
When creating a model with invalid fields, we get an HTTP 422 which is correct but
this warning is logged:

app/controllers/admin/base_controller.rb:13:in `authenticate'
gems/rack-3.2.1/lib/rack/utils.rb:584: warning: Status code :unprocessable_entity is deprecated and will be removed in a future version of Rack. Please use :unprocessable_content instead.

This was deprecated in Rails 7.1